library switch: fix switch on mobile
