+++ projects/admin/src/app/menu/menu-mobile/sub-menu/sub-menu.component.html
@@ -1,6 +1,6 @@
 <!--
   RERO ILS UI
-  Copyright (C) 2020 RERO
+  Copyright (C) 2020-2023 RERO
 
   This program is free software: you can redistribute it and/or modify
   it under the terms of the GNU Affero General Public License as published by
@@ -30,6 +30,7 @@
             [attr.target]="subItem.getAttribute('target')"
             [routerLink]="subItem.getRouterLink()"
             [queryParams]="subItem.getQueryParams() ? subItem.getQueryParams() : {}"
+            (click)="doClick(subItem)"
           >
             <i *ngIf="subItem.hasExtra('iconClass')" [ngClass]="subItem.getExtra('iconClass')"></i>
             {{ subItem.getName() }}
